CrisisCompassAPI/requirements.txt
flask
torch
transformers
spacy
geopy
python-dotenv
supabase
requests
